fix movie view model and move api as env
