+++ app/src/main/java/Network/RetrofitClient.kt
@@ -1,23 +1,23 @@
-package com.example.moviemues.Network
+package com.example.moviemuse.Network
 
-import com.example.moviemues.BuildConfig
+import com.example.moviemuse.BuildConfig
 import retrofit2.Retrofit
 import retrofit2.converter.gson.GsonConverterFactory
 import okhttp3.Interceptor
 import okhttp3.OkHttpClient
 import okhttp3.Request
+import okhttp3.logging.HttpLoggingInterceptor
 import java.util.concurrent.TimeUnit
 
 object RetrofitClient {
-    private const val BASE_URL = "https://api.themoviedb.org/3/"
+    private const val BASE_URL = BuildConfig.TMDB_BASE_URL
+
+    private val loggingInterceptor = HttpLoggingInterceptor().apply {
+        level = HttpLoggingInterceptor.Level.BODY
+    }
 
     private val client = OkHttpClient.Builder()
-        .addInterceptor { chain ->
-            val request = chain.request().newBuilder()
-                .addHeader("Authorization", "Bearer ${BuildConfig.TMDB_API_KEY}")
-                .build()
-            chain.proceed(request)
-        }
+//        .addInterceptor(loggingInterceptor) // Remove Comment to logs API requests
         .connectTimeout(30, TimeUnit.SECONDS)
         .writeTimeout(30, TimeUnit.SECONDS)
         .readTimeout(30, TimeUnit.SECONDS)
