+++ app/src/main/java/com/example/moviemuse/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.moviemues
+package com.example.moviemuse
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -18,8 +18,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.example.moviemues.ui.theme.MovieMuseTheme
-import com.example.moviemuse.MovieListScreen
+import com.example.moviemuse.ui.theme.MovieMuseTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
